+++ src/webparts/workMinders/components/contentView/taskItem/TaskItem.module.scss
